# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Open-Kanban API
  description: |
    Open-Kanban is a self-hosted kanban board API server.
    
    ## Authentication
    
    This API uses Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    Some endpoints (marked as public) do not require authentication.
    
    ## Signature Verification (IMPORTANT SECURITY INFORMATION)
    
    Protected endpoints support HMAC-SHA256 signature verification for enhanced security. However, **signature verification is DISABLED by default**.
    
    ### Security Warning
    
    When signature verification is disabled (the default), protected endpoints will accept requests without signature verification, even if they are marked as requiring signature verification in this documentation.
    
    ### To Enable Signature Verification
    
    Set the following environment variables:
    
    1. `SIGNATURE_ENABLED=1` - Enable signature verification
    2. `SIGNATURE_SECRETS=access-key-1:secret-key-1,access-key-2:secret-key-2` - Configure access keys and their corresponding secrets (comma-separated key:value pairs)
    
    ### How Signature Verification Works
    
    When enabled, protected endpoints require the following headers:
    
    - `X-Signature`: HMAC-SHA256 signature (hex-encoded)
    - `X-Timestamp`: Unix timestamp of the request (must be within 5 minutes of server time)
...
    
    ### Signature Generation
    
...
    
...
servers:
...
paths:
...
  /auth/activities:
    get:
      summary: Get activities
      description: Get activity log
      operationId: getActivities
      tags:
        - Authentication
      security:
        - BearerAuth: []
        - SignatureAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Activity log
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SignatureAuth:
      type: apiKey
      in: header
      name: X-Signature
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Not logged in
...